refact: renaming extensions and adjusting instrumented tests
