+++ buildSrc/src/main/java/com/picpay/buildSrc/Versions.kt
@@ -43,6 +43,6 @@ object Versions {
     const val mockito = "2.27.0"
     const val mockitoInline = "2.8.47"
     const val mockitoKotlin = "2.1.0"
-    const val testRunner = "1.1.1"
-    const val espresso = "3.1.1"
+    const val testRunner = "1.3.0"
+    const val espresso = "3.3.0"
 }
